[FIX] product_replace: inform about `create_uid` not set on records
e.g. after deleting a user.
Instead of silently ignoring it and proceeding, we raise a warning so
admin can init create_uid to the correct values. Deleting a user is not
a normal way of removing it, inactivating the user should be used instead.

used to cause this error in orm_enhanced.group_by():
TypeError: unhashable type: 'list'